backend/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
django-json-widget==1.1.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
django-timestampedmodel==0.1.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn[gevent]
openpyxl==3.1.2
pandas==2.2.1
pre-commit==3.6.2
pylint==3.1.0
pylint-django==2.5.5
pytest==8.1.1
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-xdist==3.5.0
python-decouple==3.8
